improve latency benchmark: don't exceed submitter limits, sort by nonce, don't pull in submitter business logic
